create utils.py and remove others
