+++ src/sdk/pynni/nni/hyperopt_tuner/hyperopt_tuner.py
@@ -29,6 +29,7 @@ import numpy as np
 
 import hyperopt as hp
 from nni.tuner import Tuner
+from nni.utils import *
 
 logger = logging.getLogger('hyperopt_AutoML')
 
@@ -241,7 +242,7 @@ class HyperoptTuner(Tuner):
             if value is dict, it should have "default" key.
             value is final metrics of the trial.
         """
-        reward = self.extract_scalar_reward(value)
+        reward = extract_scalar_reward(value)
         # restore the paramsters contains '_index'
         if parameter_id not in self.total_data:
             raise RuntimeError('Received parameter_id not in total_data.')
