+++ src/sdk/pynni/nni/tuner.py
@@ -103,12 +103,3 @@ class Tuner(Recoverable):
 
     def _on_error(self):
         pass
-
-    def extract_scalar_reward(self, value, scalar_key='default'):
-        if isinstance(value, float) or isinstance(value, int):
-            reward = value
-        elif isinstance(value, dict) and scalar_key in value and isinstance(value[scalar_key], (float, int)):
-            reward = value[scalar_key]
-        else:
-            raise RuntimeError('Incorrect final result: the final result for %s should be float/int, or a dict which has a key named "default" whose value is float/int.' % str(self.__class__)) 
-        return reward
